Use <T>::BITS if no deku `bits` attribute
Much like the deku `bits` attribute, make the `bits` attribute in BinLex
optional and use the <T>::Bits if not supplied.

This had several side effects of having to use more quote! for some
actual parsing of the Rust AST instead of generating just a string. That
is most of this code change.

Removed optional `bits` field in example `main.rs` and added test.

Closes #3

Signed-off-by: wcampbell <wcampbell1995@gmail.com>